Update
- Added some flexural buckling calculations on the EN_tools.py

+++ closed/parametric_project.py
@@ -3,18 +3,15 @@ from itertools import product
 from shutil import rmtree
 from os import path
 
+#### INPUT ####
 # Import the method to be run parametrically
-import cylinder
-
-
-session.journalOptions.setValues(replayGeometry=COORDINATE, recoverGeometry=COORDINATE)
+import closed_polygons
 
 # Give a project name
-prj_name = 'cylinder'
-try:
-    prj_name = float(sys.argv[-1])
-except:
-    pass
+prj_name = 'specimens'
+
+# Remove folders after execution? (keep only the output file)
+remove_folders = False
 
 # Define a list of lists of input values.
 # The parrent list has length of the number of parameters that vary
@@ -22,11 +19,12 @@ except:
 # A list of all possible combinations is stored in "combinations"
 combinations = list(
     product(
-        range(6, 31, 2), 
-        range(27, 55, 3),
-        [6., 16.]
+        [12, 18, 24], 
+        [30, 40, 50]
         )
     )
+
+#### END INPUT ####
 
 # Open a file to collect the results
 out_file = open('./'+prj_name+'_info.dat', 'a')
@@ -48,17 +46,21 @@ for parameter in combinations:
     # The function to be run for the full factorial parametric is called here
     print('Running job: ' + job_ID)
     try:
-        cylinder_thickness = cylinder.polygon_2_cylinder(
-            parameter[0],
-            250.,
-            parameter[1],
-            381.
+        radius = closed_polygons.class_2_radius(
+            n_sides = parameter[0],
+            p_classification = parameter[1],
+            thickness = 3,
+            f_yield = 420.
             )
-        job_return = cylinder.modeler(
-            shell_thickness = cylinder_thickness,
-            n_of_waves = parameter[0] / 2,
-            m_of_waves = parameter[0] / 2,
-            u_max = parameter[2]/1000,
+        
+        job_return = closed_polygons.modeler(
+            n_sides = parameter[0], 
+            r_circle = radius, 
+            p_classification = parameter[1], 
+            n_of_waves = parameter[0] / 2, 
+            m_of_waves = parameter[0] / 2, 
+            u_max = 0.00667, 
+            f_yield = 420., 
             IDstring = job_ID, 
             proj_name = prj_name, 
             )
@@ -68,13 +70,16 @@ for parameter in combinations:
         
         # Write each returned string to the file separated by newlines
         job_return = str(job_return)
-        out_file.write(job_ID + " : " + job_return + "\n")
+        out_file.write(job_ID + ", " + job_return + "\n")
         
     except:
         print('Problem while executing job: '+ job_ID)
         print('Job is canceled and the folder is deleted. See log file (to be writen)')
         Mdb()
         os.chdir('../')
+    
+    # Remove job's folder (only the output information is kept)
+    if remove_folders is True:
         rmtree(job_ID)
 
 out_file.close()
